No error limit by default.
There was no other way to set this (-e=-1 didn't seem to work).  Setting this allows using "-EP -eq" to generate as much as possible code without failing on stuff not related to preprocessor.  Potentially this is wrong and needs to be fixed to fail if preprocessor macros are bad while silently ignoring other errors.